requirements.txt
einops==0.8.0
matplotlib==3.9.1
numpy==2.0.1
opencv_python_headless==4.8.1.78
Pillow==10.1.0
Pillow==10.4.0
timm==1.0.8
torch==2.0.0
torchvision==0.15.1
